+++ src/Component/categories.jsx
@@ -1,13 +1,10 @@
-/* eslint-disable */
-const Categories = () =>{
-    return(
-     <div className="categries-container">
-        <ul className="UL-categries">
-            <h2>Underconstruction</h2>
-            <li className="fiction">Fiction</li>
-            <li className="nonfiction">Nonfiction</li>
-        </ul>
-     </div>   
-    )
-}
-export default Categories
+const Categories = () => (
+  <div className="categries-container">
+    <ul className="UL-categries">
+      <h2>Underconstruction</h2>
+      <li className="fiction">Fiction</li>
+      <li className="nonfiction">Nonfiction</li>
+    </ul>
+  </div>
+);
+export default Categories;
